fix: remove motion, replace button, fix indent
