--- app/components/Arena/eventData.ts
@@ -1,42 +1,0 @@
-const eventData = [
-    {
-        title: "Intania Hackathon",
-        description:
-            "เตรียมตัวพบกับ Intania Hackathon​ - งานที่จะสร้าง บรรยากาศที่ทำให้ทุกคนได้สร้างสิ่งใหม่ ๆ แก้ไขปัญหาใหม่ ๆ ที่สร้าง Impact ได้จริงไปด้วยกัน",
-        image: "/arena/ih.png",
-        links: [
-            { name: "ชมถ่ายทอดสด", url: "/" },
-            { name: "ติดตาม", url: "/" },
-        ],
-    },
-    {
-        title: "Integration Bee Challenge",
-        description:
-            "เตรียมตัวพบกับ Intania Hackathon​ - งานที่จะสร้าง บรรยากาศที่ทำให้ทุกคนได้สร้างสิ่งใหม่ ๆ แก้ไขปัญหาใหม่ ๆ ที่สร้าง Impact ได้จริงไปด้วยกัน",
-        image: "/arena/ib.png",
-        links: [
-            { name: "สมัคร", url: "/" },
-            { name: "อ่านต่อ", url: "/" },
-        ],
-    },
-    {
-        title: "Intania Innovation Contest",
-        description:
-            "การแข่งขันนวัตกรรมนักเรียนระดับม้ธยมศึกษาตอนปลาย ในด้านวิศวกรรมศาสตร์ประจำปี 2568 ในธีม “Innovation towards Global Impact”",
-        image: "/arena/iic.png",
-        links: [
-            { name: "อ่านต่อ", url: "/" },
-        ],
-    },
-    {
-        title: "Intania Metaverse Exhibition",
-        description:
-            "Pitching รอบไฟนอลของ Thailand Metaverse Hackathon and Exhibition ครั้งที่ 1 และงานจัดแสดง Metaverse โดยบริษัทและหน่วยงานต่าง ๆ",
-        image: "/arena/ime.png",
-        links: [
-            { name: "อ่านต่อ", url: "/" },
-        ],
-    },
-];
-
-export default eventData;
